refactor(remote-config): split modular and namespaced entrypoints
