+++ packages/remote-config/lib/modular.d.ts
@@ -1,0 +1,1 @@
+export * from './modular/index';
